migrated gui

+++ core/components/rectangle.py
@@ -1,7 +1,7 @@
 class Rectangle:
     """A rectangle supporting basic collision detection."""
 
-    def __init__(self, pos: tuple[float, float], width: int, height: int):
+    def __init__(self, pos: tuple[int, int], width: int, height: int):
         self.pos = pos
         self.width, self.height = width, height
 
@@ -13,18 +13,18 @@ class Rectangle:
         """Gets the height of the image."""
         return self.height
 
-    def get_position(self) -> tuple[float, float]:
+    def get_position(self) -> tuple[int, int]:
         """Gets the position of the rectangle."""
         return self.pos
 
-    def set_position(self, pos: tuple[float, float]):
+    def set_position(self, pos: tuple[int, int]):
         """Sets the position of the rectangle."""
         self.pos = pos
 
-    def add_position(self, pos: tuple[float, float]):
+    def add_position(self, pos: tuple[int, int]):
         """Adds to the position of the rectangle."""
         self.pos = (self.pos[0] + pos[0], self.pos[1] + pos[1])
 
-    def collides_with_point(self, pos: tuple[float, float]) -> bool:
+    def collides_with_point(self, pos: tuple[int, int]) -> bool:
         """Checks if a point's position is within the rectangle."""
         return self.pos[0] <= pos[0] <= self.pos[0] + self.width and self.pos[1] <= pos[1] <= self.pos[1] + self.height
